+++ src/server/autoCompletion.ts
@@ -255,7 +255,8 @@ export class AutoCompletion {
         return symList.length > 0 ? symList : null;
     }
 
-    /* 搜索模块名
+    /**
+     * 搜索模块名
      * 正常情况下，声明一个模块都会产生一个符号
      * 但table.empty = function() ... end这种扩展标准库或者C++导出的模块时就没有
      * 所以这里特殊处理